model/requirements.txt
# Base ------------------------------------------------------------------------
gitpython
ipython  # interactive notebook
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow==9.0.1
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0
comet_ml==3.31.20
jupyter==1.0.0

# API -------------------------------------------------------------------------
flask==2.0.3
gunicorn==20.1.0
pytest==7.1.1
redis==4.1.4
werkzeug==2.0.3

# Model -----------------------------------------------------------------------
pytest==7.1.1
protobuf==3.20.0
boto3==1.24.7

# Logging ---------------------------------------------------------------------
tensorboard>=2.4.1
# clearml>=1.2.0
# comet

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
